build: full renovation of dependencies to remove vulnerabilities
1. Also bumped the OpenAPI spec file to v3 from v2 because I it seems
that it was already bumped earlier by whoever generated the code currently
under version control but they did not appear to have bumped the spec file.
(I could tell this because the generated code had v3 in the comments
which flicked back to v2 when I ran the generator script initially)
2. Added the openapi generator CLI npm dependency so now the downloading
of the .jar file happens automatically.
3. The tests are passing but the coverage script runs out of heap memory
during execution so I couldn't verify that that one also works (or
to put it another way, I have verified that it does not)

Signed-off-by: Peter Somogyvari <peter.somogyvari@accenture.com>